poem part 1 into part 2 - coming to term | terms
In the section of the poem I have written so far, I have covered what I mean by coming to term (the first iteration of this episode's theme/title/first line). I am trying to say that coming to term is a way of gathering everything from the previous milestone/term point and reflecting on the meaning while also considering the implications for the future -- I really relate to this feeling seeing that I am at the end of my MA and creating this (and many other) project(s) as a part of that.  
  
In the final line group before the horizontal rule (which indicates the episodal shift from poetry to analysis) I have started poetically summarizing what I mean by the second iteration of this episode's theme/title/first line - coming to terms. This section of the poem will work through some of my thoughts regarding the recent twitter debates regarding the usefulness of computational literary studies. I hope to give some sense of the labor and the add my point that there is value in the act even if the results or method are skewed. I think the humanities are closer than ever to hacking something the hard sciences and data scientists have been trying to do for years... a better understanding of the human psyche from the minds of those, past and present, able to record thought in words.  Still grappling... see you in the next commit!